pyproject.toml
[project]
name = "mcp-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.2",
    "dotenv>=0.9.9",
    "fastmcp>=2.14.5",
    "jsonschema>=4.26.0",
    "mcp>=1.26.0",
    "openai>=2.21.0",
]

[tool.uv]
package = true

[project.scripts]
mcp_eval = "mcp_evaluation.main:main"